+++ src/clj_watson/controller/output.clj
@@ -17,13 +17,13 @@
     (println (logic.template/generate {:vulnerable-dependencies dependencies} template))))
 
 (defmethod ^:private generate* :json [dependencies & _]
-  (-> dependencies json/generate-string pprint/pprint))
+  (-> dependencies (json/generate-string {:pretty true}) println))
 
 (defmethod ^:private generate* :edn [dependencies & _]
   (pprint/pprint dependencies))
 
 (defmethod ^:private generate* :sarif [dependencies deps-edn-path & _]
-  (-> dependencies (logic.sarif/generate deps-edn-path) json/generate-string println))
+  (-> dependencies (logic.sarif/generate deps-edn-path) (json/generate-string {:pretty true}) println))
 
 (defn generate [dependencies deps-edn-path kind]
   (generate* dependencies deps-edn-path kind))